requirements.txt
datasets>=2.19.0
huggingface_hub>=0.24.0
transformers>=4.45.0
accelerate>=0.33.0
torch>=2.1.0
opencv-python-headless>=4.8.0
scikit-learn>=1.3.0
Pillow>=10.0.0
wandb>=0.17.0
matplotlib>=3.8.0
